+++ llvm/lib/Transforms/Utils/Debugify.cpp
@@ -147,7 +147,7 @@ bool llvm::applyDebugifyMetadata(
     for (BasicBlock &BB : F) {
       // Attach debug locations.
       for (Instruction &I : BB) {
-        uint64_t AtomGroup = ApplyAtomGroups ? NextLine : 0;
+        uint32_t AtomGroup = ApplyAtomGroups ? NextLine : 0;
         uint8_t AtomRank = ApplyAtomGroups ? 1 : 0;
         uint64_t Line = NextLine++;
         I.setDebugLoc(DILocation::get(Ctx, Line, 1, SP, nullptr, false,
